moved and cleaned MakeALogo w/@dcoopersmith
